+++ src/Channel.test.ts
@@ -1,6 +1,11 @@
 import http from "http";
 import EventSource from "eventsource";
-import {createServer, closeServer, getUrl} from "./lib/testUtils";
+import {
+	createServer,
+	closeServer,
+	getUrl,
+	waitForConnect,
+} from "./lib/testUtils";
 import Session from "./Session";
 import Channel from "./Channel";
 
@@ -33,22 +38,22 @@ describe("construction", () => {
 
 describe("registering", () => {
 	it("can register and store an active session", (done) => {
-		server.on("request", (req, res) => {
+		server.on("request", async (req, res) => {
 			const session = new Session(req, res);
 
-			session.on("connected", () => {
-				const channel = new Channel();
+			await waitForConnect(session);
 
-				expect(channel.sessionCount).toBe(0);
-				expect(channel.activeSessions).toEqual([]);
+			const channel = new Channel();
 
-				channel.register(session);
+			expect(channel.sessionCount).toBe(0);
+			expect(channel.activeSessions).toEqual([]);
 
-				expect(channel.sessionCount).toBe(1);
-				expect(channel.activeSessions).toEqual([session]);
+			channel.register(session);
 
-				done();
-			});
+			expect(channel.sessionCount).toBe(1);
+			expect(channel.activeSessions).toEqual([session]);
+
+			done();
 		});
 
 		eventsource = new EventSource(url);
@@ -80,7 +85,7 @@ describe("registering", () => {
 		server.on("request", async (req, res) => {
 			const session = new Session(req, res);
 
-			await new Promise((resolve) => session.on("connected", resolve));
+			await waitForConnect(session);
 
 			channel.register(session);
 
@@ -95,22 +100,22 @@ describe("registering", () => {
 	it("removes a session from the active sessions after deregistering it", (done) => {
 		const channel = new Channel();
 
-		server.on("request", (req, res) => {
+		server.on("request", async (req, res) => {
 			const session = new Session(req, res);
 
-			session.on("connected", () => {
-				channel.register(session);
+			await waitForConnect(session);
 
-				expect(channel.activeSessions).toContain(session);
-				expect(channel.sessionCount).toBe(1);
+			channel.register(session);
 
-				channel.deregister(session);
+			expect(channel.activeSessions).toContain(session);
+			expect(channel.sessionCount).toBe(1);
 
-				expect(channel.activeSessions).not.toContain(session);
-				expect(channel.sessionCount).toBe(0);
+			channel.deregister(session);
 
-				done();
-			});
+			expect(channel.activeSessions).not.toContain(session);
+			expect(channel.sessionCount).toBe(0);
+
+			done();
 		});
 
 		eventsource = new EventSource(url);
@@ -128,7 +133,7 @@ describe("registering", () => {
 		server.on("request", async (req, res) => {
 			const session = new Session(req, res);
 
-			await new Promise((resolve) => session.on("connected", resolve));
+			await waitForConnect(session);
 
 			channel.register(session).deregister(session);
 
@@ -147,13 +152,15 @@ describe("registering", () => {
 		server.on("request", async (req, res) => {
 			const session = new Session(req, res);
 
-			await new Promise((resolve) => session.on("connected", resolve));
+			await waitForConnect(session);
 
 			channel.register(session);
 
 			const deregister = jest.spyOn(channel, "deregister");
 
-			await new Promise((resolve) => session.on("disconnected", resolve));
+			await new Promise<void>((resolve) =>
+				session.on("disconnected", resolve)
+			);
 
 			expect(deregister).toHaveBeenCalledWith(session);
 			expect(channel.activeSessions).not.toContain(session);
@@ -179,11 +186,13 @@ describe("registering", () => {
 		server.on("request", async (req, res) => {
 			const session = new Session(req, res);
 
-			await new Promise((resolve) => session.on("connected", resolve));
+			await waitForConnect(session);
 
 			channel.register(session);
 
-			await new Promise((resolve) => session.on("disconnected", resolve));
+			await new Promise<void>((resolve) =>
+				session.on("disconnected", resolve)
+			);
 
 			expect(callback).toHaveBeenCalledWith(session);
 
@@ -209,7 +218,7 @@ describe("broadcasting", () => {
 
 			const push = jest.spyOn(session, "push");
 
-			await new Promise((resolve) => session.on("connected", resolve));
+			await waitForConnect(session);
 
 			channel.register(session);
 
@@ -233,7 +242,7 @@ describe("broadcasting", () => {
 		server.on("request", async (req, res) => {
 			const session = new Session(req, res);
 
-			await new Promise((resolve) => session.on("connected", resolve));
+			await waitForConnect(session);
 
 			channel.register(session);
 
@@ -247,7 +256,7 @@ describe("broadcasting", () => {
 		eventsource = new EventSource(url);
 	});
 
-	it("can filter sessions when broadcasting", async (done) => {
+	it("can filter sessions when broadcasting", (done) => {
 		const channel = new Channel();
 
 		const sessionPushMocks: jest.SpyInstance[] = [];
@@ -255,7 +264,7 @@ describe("broadcasting", () => {
 		server.on("request", async (req, res) => {
 			const session = new Session(req, res);
 
-			await new Promise((resolve) => session.on("connected", resolve));
+			await waitForConnect(session);
 
 			sessionPushMocks.push(jest.spyOn(session, "push"));