+++ src/Session.ts
@@ -1,7 +1,7 @@
-import EventEmitter from "events";
 import {randomBytes} from "crypto";
 import {Readable} from "stream";
 import {IncomingMessage, ServerResponse, OutgoingHttpHeaders} from "http";
+import {TypedEmitter, EventMap} from "./lib/TypedEmitter";
 import serialize, {SerializerFunction} from "./lib/serialize";
 import sanitize, {SanitizerFunction} from "./lib/sanitize";
 
@@ -90,6 +90,11 @@ interface IterateOptions {
 	eventName?: string;
 }
 
+interface Events extends EventMap {
+	connected: () => void;
+	disconnected: () => void;
+}
+
 /**
  * A Session represents an open connection between the server and a client.
  *
@@ -102,7 +107,9 @@ interface IterateOptions {
  * @param res - The Node HTTP {@link https://nodejs.org/api/http.html#http_class_http_serverresponse | IncomingMessage} object.
  * @param options - Options given to the session instance.
  */
-class Session extends EventEmitter {
+class Session<
+	State extends Record<string, unknown> = Record<string, unknown>
+> extends TypedEmitter<Events> {
 	/**
 	 * The last ID sent to the client.
 	 * This is initialized to the last event ID given by the user, and otherwise is equal to the last number given to the `.id` method.
@@ -122,7 +129,7 @@ class Session extends EventEmitter {
 	 * Custom state for this session.
 	 * Use this object to safely store information related to the session and user.
 	 */
-	state: Record<string, unknown> = {};
+	state = {} as State;
 
 	private req: IncomingMessage;
 	private res: ServerResponse;